requirements.txt
asgiref==3.8.1
Django==5.1.6
django-debug-toolbar==5.0.1
django-filter==24.3
djangorestframework==3.15.2
Markdown==3.7
pillow==11.1.0
sqlparse==0.5.3
